requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-docx
azure-storage-blob
pydantic